requirements.txt
paramiko==2.7.2
gitpython==3.1.24
